+++ src/http.rs
@@ -16,7 +16,7 @@
 
 use std::{borrow::Cow, net::SocketAddr};
 
-use check_if_email_exists::{check_email, CheckEmailInput};
+use check_if_email_exists::{check_email, CheckEmailInput, CheckEmailInputProxy};
 use hyper::service::{make_service_fn, service_fn};
 use hyper::{Body, Method, Request, Response, Server, StatusCode};
 use serde::{Deserialize, Serialize};
@@ -25,6 +25,10 @@ use crate::CONF;
 
 /// JSON Request from POST /
 #[derive(Debug, Deserialize, Serialize)]
+#[deprecated(
+	since = "0.8.24",
+	note = "The HTTP server will be removed from the CLI, please use https://github.com/reacherhq/backend instead"
+)]
 pub struct PostReqBody {
 	from_email: Option<String>,
 	hello_name: Option<String>,
@@ -35,9 +39,14 @@ pub struct PostReqBody {
 
 /// Error Response from POST /
 #[derive(Debug, Deserialize, Serialize)]
+#[deprecated(
+	since = "0.8.24",
+	note = "The HTTP server will be removed from the CLI, please use https://github.com/reacherhq/backend instead"
+)]
 pub struct PostResError {
 	error: String,
 }
+
 async fn req_handler(req: Request<Body>) -> Result<Response<Body>, hyper::Error> {
 	match (req.method(), req.uri().path()) {
 		// Serve some instructions at /
@@ -63,14 +72,14 @@ async fn req_handler(req: Request<Body>) -> Result<Response<Body>, hyper::Error>
 			// Create EmailInput from body
 			let mut input = CheckEmailInput::new(body.to_emails);
 			input
-				.from_email(body.from_email.unwrap_or_else(|| CONF.from_email.clone()))
-				.hello_name(body.hello_name.unwrap_or_else(|| CONF.hello_name.clone()))
-				.yahoo_use_api(CONF.yahoo_use_api);
+				.set_from_email(body.from_email.unwrap_or_else(|| CONF.from_email.clone()))
+				.set_hello_name(body.hello_name.unwrap_or_else(|| CONF.hello_name.clone()))
+				.set_yahoo_use_api(CONF.yahoo_use_api);
 			if let Some(proxy_host) = body.proxy_host.map(Cow::Owned).or_else(|| CONF.proxy_host.as_ref().map(Cow::Borrowed)) {
-				input.proxy(
-					proxy_host.into_owned(),
-					body.proxy_port.unwrap_or(CONF.proxy_port),
-				);
+				input.set_proxy(CheckEmailInputProxy {
+					host:proxy_host.into_owned(),
+					port: body.proxy_port.unwrap_or(CONF.proxy_port)
+				});
 			}
 
 			let body = check_email(&input).await;
@@ -99,9 +108,15 @@ async fn req_handler(req: Request<Body>) -> Result<Response<Body>, hyper::Error>
 	}
 }
 
+#[deprecated(
+	since = "0.8.24",
+	note = "The HTTP server will be removed from the CLI, please use https://github.com/reacherhq/backend instead"
+)]
 pub async fn run<A: Into<SocketAddr>>(
 	addr: A,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+	println!("WARNING: The HTTP server is deprecated, and will be removed in v0.9.0. Please see https://github.com/reacherhq/backend for a replacement.");
+
 	let service = make_service_fn(|_| async { Ok::<_, hyper::Error>(service_fn(req_handler)) });
 
 	let addr = addr.into();