+++ src/main.rs
@@ -18,7 +18,7 @@ mod http;
 
 use std::net::IpAddr;
 
-use check_if_email_exists::{check_email, CheckEmailInput};
+use check_if_email_exists::{check_email, CheckEmailInput, CheckEmailInputProxy};
 use clap::Clap;
 use once_cell::sync::Lazy;
 
@@ -51,19 +51,34 @@ pub struct Cli {
 	/// The email to check.
 	pub to_email: Option<String>,
 
-	/// Runs a HTTP server.
+	/// DEPRECATED. Runs a HTTP server.
+	/// This option will be removed in v0.9.0.
 	#[clap(long)]
+	#[deprecated(
+		since = "0.8.24",
+		note = "The HTTP server will be removed from the CLI, please use https://github.com/reacherhq/backend instead"
+	)]
 	pub http: bool,
 
-	/// Sets the host IP address on which the HTTP server should bind.
+	/// DEPRECATED. Sets the host IP address on which the HTTP server should bind.
 	/// Only used when `--http` flag is on.
+	/// This option will be removed in v0.9.0.
 	#[clap(long, env = "HOST", default_value = "127.0.0.1")]
+	#[deprecated(
+		since = "0.8.24",
+		note = "The HTTP server will be removed from the CLI, please use https://github.com/reacherhq/backend instead"
+	)]
 	pub http_host: IpAddr,
 
-	/// Sets the port on which the HTTP server should bind.
+	/// DEPRECATED. Sets the port on which the HTTP server should bind.
 	/// Only used when `--http` flag is on.
 	/// If not set, then it will use $PORT, or default to 3000.
+	///  This option will be removed in v0.9.0.
 	#[clap(long, env = "PORT", default_value = "3000")]
+	#[deprecated(
+		since = "0.8.24",
+		note = "The HTTP server will be removed from the CLI, please use https://github.com/reacherhq/backend instead"
+	)]
 	pub http_port: u16,
 }
 
@@ -77,11 +92,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
 	if let Some(to_email) = &CONF.to_email {
 		let mut input = CheckEmailInput::new(vec![to_email.clone()]);
 		input
-			.from_email(CONF.from_email.clone())
-			.hello_name(CONF.hello_name.clone())
-			.yahoo_use_api(CONF.yahoo_use_api);
+			.set_from_email(CONF.from_email.clone())
+			.set_hello_name(CONF.hello_name.clone())
+			.set_yahoo_use_api(CONF.yahoo_use_api);
 		if let Some(proxy_host) = &CONF.proxy_host {
-			input.proxy(proxy_host.clone(), CONF.proxy_port);
+			input.set_proxy(CheckEmailInputProxy {
+				host: proxy_host.clone(),
+				port: CONF.proxy_port,
+			});
 		}
 
 		let result = check_email(&input).await;
@@ -96,7 +114,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
 		};
 	}
 
-	// Run the web server if flag is on
+	// Run the web server if --http flag is on.
 	if CONF.http {
 		http::run((CONF.http_host, CONF.http_port)).await?;
 	}